requirements.txt
asgiref==3.7.2
Django==5.0
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-nested-routers==0.93.5
mysqlclient==2.2.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
